+++ services/login/src/factory/challenge_response_session_factory.cpp
@@ -1,11 +1,11 @@
-#include <drogon_extended/json_mapper/json_converter.hpp>
+#include "drogon_extended/json_mapper/json_converter.hpp"
 
-#include "include/controller/factory/challenge_response_session_factory.hpp"
+#include "include/factory/challenge_response_session_factory.hpp"
 
 ChallengeResponseSession ChallengeResponseSessionFactory::from_json(const std::shared_ptr<Json::Value>& jsonPtr)
 {
     auto sessionId = JsonConverter::check(jsonPtr, "session_id").asString();
-    auto challengeResponse = JsonConverter::check(jsonPtr, "challenge_response").asInt();
+    auto challengeResponse = JsonConverter::check(jsonPtr, "challenge_response").asInt64();
 
     return {sessionId, challengeResponse};
 }
